CORE-1329 Refactor App Launch FileInput param API validation
Refactored file/folder input validation into common component so that
Folder inputs can be validated as well.
